Add some custom templates
These are custom additions that were made by Punch on his
website, but the references to these are in the theme.
Thus to prevent any template errors a copy of these is
added to the theme.

These are quite generic so this addition seems an acceptable
change to the theme.

It can be improved though by making some of these things
configurable through config.toml